Fixed another BlockDB-related crash on startup.

+++ fCraft/World/BlockDB.cs
@@ -384,7 +384,8 @@ namespace fCraft {
         internal void Clear() {
             lock( SyncRoot ) {
                 CacheClear();
-                File.Delete( FileName );
+                if( File.Exists( FileName ) )
+                    File.Delete( FileName );
             }
         }
